requirements.txt
asyncio==3.4.3
joblib==1.3.2
keras==2.14.0
matplotlib==3.8.0
numpy==1.26.0
pandas==2.1.1
PyQt5==5.15.9
python-chess==1.999
scikit-learn==1.3.1
scipy==1.11.3
tabulate==0.9.0
tensorflow==2.14.0
